rhestless_events_management/requirements.txt
asgiref==3.5.2
Django==4.1
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.2.1
screen==1.0.1
Serializer==0.2.1
simplejson==3.17.6
sqlparse==0.4.2
tzdata==2022.2
xlwt==1.3.0
